going to execute CRED operation

+++ app.py
@@ -44,9 +44,11 @@ def update():
 @app.route('/delete/<int:sno>')
 def delete(sno):
     todo = Todo.query.filter_by(sno=sno).first()
-    db.session.delete(todo)
-    db.session.commit()
+    if todo:
+        db.session.delete(todo)
+        db.session.commit()
     return redirect("/")
+
 
 
 if __name__ == '__main__':
